+++ src/mockVscode.spec.ts
@@ -157,7 +157,8 @@ export let vscode = {
             return {
                 clear: () => { },
                 text: '',
-                show: () => { }
+                show: () => { },
+                dispose: () => { }
             };
         },
         createQuickPick: () => {
@@ -206,6 +207,12 @@ export let vscode = {
             } as OutputChannel;
         },
         registerTreeDataProvider: function(viewId: string, treeDataProvider: TreeDataProvider<any>) { },
+        showInformationMessage: function(message: string) {
+
+        },
+        showWarningMessage: function(message: string) {
+
+        },
         showErrorMessage: function(message: string) {
 
         },
@@ -261,32 +268,32 @@ export let vscode = {
             this.line = line;
             this.character = character;
         }
-        private line: number;
-        private character: number;
+        public line: number;
+        public character: number;
     },
     ParameterInformation: class {
         constructor(label: string, documentation?: any) {
             this.label = label;
             this.documentation = documentation;
         }
-        private label: string;
-        private documentation: any;
+        public label: string;
+        public documentation: any;
     },
     SignatureHelp: class {
         constructor() {
             this.signatures = [];
         }
-        private signatures: any[];
-        private activeParameter: number;
-        private activeSignature: number;
+        public signatures: any[];
+        public activeParameter: number;
+        public activeSignature: number;
     },
     SignatureInformation: class {
         constructor(label: string, documentation?: any) {
             this.label = label;
             this.documentation = documentation;
         }
-        private label: string;
-        private documentation: any;
+        public label: string;
+        public documentation: any;
     },
     Range: class {
         constructor(startLine: number, startCharacter: number, endLine: number, endCharacter: number) {
@@ -295,10 +302,10 @@ export let vscode = {
             this.endLine = endLine;
             this.endCharacter = endCharacter;
         }
-        private startLine: number;
-        private startCharacter: number;
-        private endLine: number;
-        private endCharacter: number;
+        public startLine: number;
+        public startCharacter: number;
+        public endLine: number;
+        public endCharacter: number;
     },
     SymbolKind: {
         File: 0,
@@ -335,7 +342,7 @@ export let vscode = {
         }
 
         private text: any;
-        private fileName: string;
+        public fileName: string;
         public getText() {
             return this.text;
         }
@@ -368,14 +375,14 @@ export let vscode = {
             this.range = range;
             this.uri = uri;
         }
-        private range: any;
-        private uri: string;
+        public range: any;
+        public uri: string;
     },
     MarkdownString: class {
         constructor(value: string = null) {
             this.value = value;
         }
-        private value: string;
+        public value: string;
     },
     ThemeColor: class { },
     Uri: {
@@ -392,7 +399,7 @@ export let vscode = {
         constructor(value: string = null) {
             this.value = value;
         }
-        private value: string;
+        public value: string;
     }
 };